+++ services/madoc-ts/src/frontend/shared/components/ViewExternalContent.tsx
@@ -1,5 +1,6 @@
 import { Preset } from '@atlas-viewer/atlas';
 import React, { useMemo } from 'react';
+import { useSiteConfiguration } from '../../site/features/SiteConfigurationContext';
 import { useContentType } from '../capture-models/plugin-api/hooks/use-content-type';
 
 export const ViewExternalContent: React.FC<{
@@ -8,6 +9,10 @@ export const ViewExternalContent: React.FC<{
   onCreated?: (runtime: Preset) => void;
   onPanInSketchMode?: () => void;
 }> = ({ target, children, onCreated, onPanInSketchMode, height = 600 }) => {
+  const {
+    project: { atlasBackground },
+  } = useSiteConfiguration();
+
   return useContentType(
     useMemo(
       () =>
@@ -23,6 +28,7 @@ export const ViewExternalContent: React.FC<{
         custom: {
           onPanInSketchMode,
           onCreateAtlas: onCreated,
+          backgroundColor: atlasBackground,
         },
       }),
       [onPanInSketchMode, onCreated, children, height]